refactor: move all PopupMessage and its function to ChatBot.tsx
